requirements.txt
keras==2.8.0
matplotlib==3.7.5
numpy==1.22.4
pandas==2.0.3
silabs_mltk==0.19.0
tensorflow==2.8.0
